src/register/requirements.txt
flask
flask-restx
flask-cors
pyyaml